python/firestore_read/google_app_engine/requirements.txt
Flask==2.1.0
google-cloud-firestore==2.5.2